requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.6
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-js-asset==3.0.1
python-decouple==3.8
sqlparse==0.5.3
